Simplify numbered_widget_item
